refactor(connections): centralize OpenWork server lifecycle
